Refactor: Modularize shared codebase into core and feature modules
