+++ feature/lightsaber/src/commonMain/kotlin/xyz/alaniz/aaron/lightsaber/feature/lightsaber/LightsaberScreen.kt
@@ -1,7 +1,5 @@
 package xyz.alaniz.aaron.lightsaber.feature.lightsaber
 
 import com.slack.circuit.runtime.screen.Screen
-import kotlinx.parcelize.Parcelize
 
-@Parcelize
-data object LightsaberScreen : Screen
+interface LightsaberScreen : Screen
