+++ feature/lightsaber/src/commonMain/kotlin/xyz/alaniz/aaron/lightsaber/feature/lightsaber/LightsaberUi.kt
@@ -41,11 +41,11 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.slack.circuit.codegen.annotations.CircuitInject
 import dev.zacsweers.metro.AppScope
-import lightsaber.shared.generated.resources.Res
-import lightsaber.shared.generated.resources.lightsaber_handle
-import lightsaber.shared.generated.resources.lightsaber_screen_lightsaber_blade
-import lightsaber.shared.generated.resources.lightsaber_screen_lightsaber_handle
-import lightsaber.shared.generated.resources.lightsaber_screen_settings_icon
+import lightsaber.feature.lightsaber.generated.resources.Res
+import lightsaber.feature.lightsaber.generated.resources.lightsaber_handle
+import lightsaber.feature.lightsaber.generated.resources.lightsaber_screen_lightsaber_blade
+import lightsaber.feature.lightsaber.generated.resources.lightsaber_screen_lightsaber_handle
+import lightsaber.feature.lightsaber.generated.resources.lightsaber_screen_settings_icon
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import xyz.alaniz.aaron.lightsaber.ui.common.LightsaberTheme
