+++ feature/settings/src/commonMain/kotlin/xyz/alaniz/aaron/lightsaber/feature/settings/SettingsScreen.kt
@@ -1,7 +1,5 @@
 package xyz.alaniz.aaron.lightsaber.feature.settings
 
 import com.slack.circuit.runtime.screen.Screen
-import kotlinx.parcelize.Parcelize
 
-@Parcelize
-data object SettingsScreen : Screen
+interface SettingsScreen : Screen
